requirements.txt
beir
sentence-transformers
faiss-cpu
rank-bm25
pyyaml
tqdm
numpy
pandas
